\name{DFresid}
\alias{DFresid}
\title{Macro to return the degrees of freedom associated with the residuals of a linear model.}
\usage{
  DFresid(lmdl = stop("lmdl not supplied"))
}
\arguments{
  \item{lmdl}{a linear model object}
}
\value{
  The degrees of freedom of the residuals of the linear
  model lmdl.
}
\description{
  Macro to return the degrees of freedom associated with
  the residuals of a linear model.
}
\author{
  Ben Veal
}

