\name{MSS}
\alias{MSS}
\title{Macro to calculate the model sum squares of a linear model}
\usage{
  MSS(lmdl = stop("lmdl not supplied"))
}
\arguments{
  \item{lmdl}{a linear model object}
}
\value{
  The model sum squares of a linear model
}
\description{
  Macro to calculate the model sum squares of a linear
  model
}
\author{
  Ben Veal
}

