\name{RSS}
\alias{RSS}
\title{Macro to calculate the residual sum squares of a linear model}
\usage{
  RSS(lmdl = stop("lmdl not supplied"))
}
\arguments{
  \item{lmdl}{a linear model object}
}
\value{
  The residual sum squares of a linear model
}
\description{
  Macro to calculate the residual sum squares of a linear
  model
}
\author{
  Ben Veal
}

