\name{TSS}
\alias{TSS}
\title{Macro to calculate the total sum squares of a linear model}
\usage{
  TSS(lmdl = stop("lmdl not supplied"))
}
\arguments{
  \item{lmdl}{a linear model object}
}
\value{
  The total sum squares of a linear model
}
\description{
  Macro to calculate the total sum squares of a linear
  model
}
\author{
  Ben Veal
}

